LayoutTests/imported/w3c/web-platform-tests/web-animations/interfaces/Animation/effect-expected.txt


PASS effect is set correctly. 
FAIL Clearing and setting Animation.effect should update the computed style of the target element assert_equals: animation no longer has an effect expected "auto" but got "100px"
